+++ src/script.rs
@@ -4,7 +4,7 @@ use std::path::Path;
 
 use crate::api::{
     base, base64_api, dotenv as api_dotenv, http as api_http, ini_parser as api_ini, io as api_io,
-    json as api_json, os as api_os, toml as api_toml, yaml as api_yaml,
+    json as api_json, os as api_os, toml as api_toml, xml as api_xml, yaml as api_yaml,
 };
 
 pub fn execute_script(file: &str, safe_mode: &bool) -> Result<()> {
@@ -33,6 +33,7 @@ pub fn execute_script(file: &str, safe_mode: &bool) -> Result<()> {
     let dapi_yaml = api_yaml::register(&lua)?;
     let dapi_ini = api_ini::register(&lua)?;
     let dapi_base64 = base64_api::register(&lua)?;
+    let dapi_xml = api_xml::register(&lua)?;
 
     let globals = lua.globals();
     let package: Table = globals.get("package")?;
@@ -74,6 +75,10 @@ pub fn execute_script(file: &str, safe_mode: &bool) -> Result<()> {
     preload.set(
         "dapi_base64",
         lua.create_function(move |_, ()| Ok(dapi_base64.clone()))?,
+    )?;
+    preload.set(
+        "dapi_xml",
+        lua.create_function(move |_, ()| Ok(dapi_xml.clone()))?,
     )?;
 
     lua.load(&script).exec()?;