+++ src/utils/mod.rs
@@ -1,5 +1,4 @@
 pub mod http_utils;
 pub mod json_utils;
 pub mod toml_utils;
-pub mod xml_utils;
 pub mod zip_utils;