+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - delered NSIS Admin Installer
 - added function to wait for a certain period of time
 - added Ini Parser
+- added Data Tests
 
 ## 0.1.11 - 05.10.2025
 - added Function to delete a directory